Brought everything forward to py3. This entailed some fairly important changes.
The biggest change behavior-wise is the use of math.fsum, which appears to have the same
properties desired from the new sum without the negative side effects, and the use of a
more straightforward product function. This changes the behavior of that function to
return the true result of multiplication rather than always returning a float.

Most other changes are superficial- changing the #! line, parens around print statements,
etc.

One doctest was changed to reflect the change to product().